+++ src/scripts/hitObjects.js
@@ -99,6 +99,15 @@ define(function(require) {
 				this.initialTime = initialTime;
 				this.lifetime = lifetime;
 			}
+		},
+		Event: class HitEvent {
+			constructor(type, score, combo, x, y) {
+				this.type = type;
+				this.score = score;
+				this.combo = combo;
+				this.x = x;
+				this.y = y;
+			}
 		}
 	};
 });